API Documentation

Portfolio analytics

Modules

OnlinePortfolioAnalytics.OnlinePortfolioAnalytics — Module

OnlinePortfolioAnalytics module aims to provide users with functionality for performing quantitative portfolio analytics via online algorithms.

MIT License

Copyright (c) 2024 FemtoTrader <femto.trader@gmail.com>

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

Asset return

OnlinePortfolioAnalytics.SimpleAssetReturn — Type

mutable struct SimpleAssetReturn{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

SimpleAssetReturn{T}(; period::Int = 1)

The SimpleAssetReturn implements asset return (simple method) calculations.

Parameters

period

Usage

Feed SimpleAssetReturn one observation at a time

julia> using OnlinePortfolioAnalytics

julia> ret = SimpleAssetReturn{Float64}()
SimpleAssetReturn: n=0 | value=missing

julia> fit!(ret, 10.0)
SimpleAssetReturn: n=1 | value=missing

julia> fit!(ret, 11.0)
SimpleAssetReturn: n=2 | value=0.1

julia> value(ret)
0.1

OnlinePortfolioAnalytics.LogAssetReturn — Type

mutable struct LogAssetReturn{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

LogAssetReturn{T}(; period::Int = 1)

The LogAssetReturn implements asset return (natural log method) calculations.

Parameters

period

Usage

Feed LogAssetReturn one observation at a time

julia> using OnlinePortfolioAnalytics

julia> ret = LogAssetReturn{Float64}()
LogAssetReturn: n=0 | value=missing

julia> fit!(ret, 10.0)
LogAssetReturn: n=1 | value=missing

julia> fit!(ret, 11.0)
LogAssetReturn: n=2 | value=0.0953102

julia> value(ret)
0.09531017980432493

Mean return

OnlineStatsBase.Mean

OnlinePortfolioAnalytics.GeometricMeanReturn — Type

mutable struct GeometricMeanReturn{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

GeometricMeanReturn{T}()

The GeometricMeanReturn type implements geometric mean returns calculations.

Cumulative return

OnlinePortfolioAnalytics.CumulativeReturn — Type

mutable struct CumulativeReturn{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

CumulativeReturn{T}()

The CumulativeReturn type implements cumulative return calculations.

Standard deviation

OnlinePortfolioAnalytics.StdDev — Type

mutable struct StdDev{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

StdDev{T}()

The StdDev type implements standard deviation calculations.

Drawdowns

OnlinePortfolioAnalytics.DrawDowns — Type

mutable struct DrawDowns{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

DrawDowns{T}()

The DrawDowns type implements drawdowns calculations (geometric method).

OnlinePortfolioAnalytics.ArithmeticDrawDowns — Type

mutable struct ArithmeticDrawDowns{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

ArithmeticDrawDowns{T}()

The ArithmeticDrawDowns type implements drawdowns calculations (arithmetic method).

Statistical moments

OnlinePortfolioAnalytics.AssetReturnMoments — Type

mutable struct AssetReturnMoments{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsMultiOutput{T}

AssetReturnMoments{T}()

The AssetReturnMoments type implements 4 first statistical moments (mean, std, skewness, kurtosis) calculations.

Sharpe ratio

OnlinePortfolioAnalytics.Sharpe — Type

mutable struct Sharpe{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

Sharpe{T}(; period=252, risk_free=0)

The Sharpe type implements sharpe ratio calculations.

Parameters

period: default is 252. Daily (252), Hourly (252*6.5), Minutely(252*6.5*60) etc...

risk_free: default is 0. Constant risk-free return throughout the period.

Sortino ratio

OnlinePortfolioAnalytics.Sortino — Type

mutable struct Sortino{T} <: OnlinePortfolioAnalytics.PortfolioAnalyticsSingleOutput{T}

Sortino{T}(; period=252, risk_free=0)

The Sortino type implements Sortino ratio calculations.

Parameters

period: default is 252. Daily (252), Hourly (252*6.5), Minutely(252*6.5*60) etc...

risk_free: default is 0. Constant risk-free return throughout the period.

Other

OnlinePortfolioAnalytics.Prod — Type

Prod(T::Type = Float64)

Track the overall product.